refactor: migrated react2-lighthouse-viewer to vitejs
