+++ packages/lighthouse-viewer/package.json
@@ -43,9 +43,7 @@
   "devDependencies": {
     "typed-query-selector": "2.6.1",
     "typescript": "4.4.4",
-    "vite": "2.7.2"
-  },
-  "dependencies": {
+    "vite": "2.7.2",
     "lighthouse": "9.2.0"
   },
   "browserslist": [
