+++ packages/react2-lighthouse-viewer/.gitignore
@@ -1,0 +1,6 @@
+node_modules
+.DS_Store
+dist
+dist-ssr
+*.local
+demo
