+++ packages/react2-lighthouse-viewer/index.html
@@ -16,8 +16,6 @@
       gtag('config', 'UA-105846650-1');
     </script>
     <link rel="stylesheet" href="https://unpkg.com/mvp.css" />
-    <script type="module" crossorigin src="/assets/index.0f896fc9.js"></script>
-    <link rel="modulepreload" href="/assets/vendor.5c4252e4.js">
   </head>
   <body>
     <header>
@@ -56,7 +54,7 @@
     </header>
     <main>
       <div id="react-app"></div>
-      
+      <script type="module" src="/src/main.tsx"></script>
     </main>
     <footer>
       <p>
