+++ packages/react2-lighthouse-viewer/lib/main.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { renderReport } from 'lighthouse-viewer';
 
 export type React2LighthouseViewerProps = {
@@ -8,7 +8,7 @@ export type React2LighthouseViewerProps = {
 export default ({ json }: React2LighthouseViewerProps): JSX.Element => {
   const [template, setTemplate] = useState('');
   useEffect(() => {
-    setTemplate(renderReport(json).outerHTML);
+    setTemplate(renderReport(json, {}).outerHTML);
   }, [json]);
 
   return <div dangerouslySetInnerHTML={{ __html: template }} />;
