+++ packages/react2-lighthouse-viewer/src/assets/report-v7.5.0.json
