+++ packages/react2-lighthouse-viewer/src/assets/report-v8.0.0.json
