+++ packages/react2-lighthouse-viewer/src/assets/report.json
