--- packages/react2-lighthouse-viewer/src/index.ts
@@ -1,3 +1,0 @@
-import React2LighthouseViewer from './components/React2LighthouseViewer';
-export { React2LighthouseViewer };
-export default React2LighthouseViewer;
