+++ packages/react2-lighthouse-viewer/src/main.tsx
@@ -1,6 +1,6 @@
 import { render } from 'react-dom';
-import reportJson from './report-v8.0.0.json';
-import React2LighthouseViewer from '..';
+import reportJson from './assets/report-v8.0.0.json';
+import React2LighthouseViewer from '../lib/main';
 import React, { FC, useEffect, useState } from 'react';
 
 const reactRootElement = document.getElementById('react-app');
