--- packages/react2-lighthouse-viewer/src/shims.d.ts
@@ -1,4 +1,0 @@
-declare module '*.html' {
-  const value: string;
-  export default value;
-}
