--- packages/react2-lighthouse-viewer/tsconfig.d.json
@@ -1,13 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "baseUrl": "src/",
-        "declaration": true,
-        "declarationDir": "dist/",
-        "emitDeclarationOnly": true,
-        "jsx": "react"
-    },
-    "include": [
-        "./src/**/*"
-    ]
-}
