+++ packages/react2-lighthouse-viewer/tsconfig.json
@@ -1,15 +1,7 @@
 {
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "strict": true,
-        "baseUrl": "src/",
-        "declaration": false,
-        "allowJs": false,
-        "jsx": "react",
-        "resolveJsonModule": true
-    },
-    "include": [
-        "./src/**/*",
-        "./demo/**/*"
-    ]
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "noEmit": true,
+  },
+  "include": ["./src", "./lib/"]
 }
